+++ feature/chat/src/main/kotlin/com/ssavice/chat/ui/ChatDivider.kt
@@ -1,0 +1,51 @@
+package com.ssavice.chat.ui
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ChatDivider(
+    message: String,
+    modifier: Modifier = Modifier,
+    color: Color = MaterialTheme.colorScheme.outlineVariant, // 기본 색상
+    textColor: Color = MaterialTheme.colorScheme.outline // 텍스트 색상
+) {
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(vertical = 16.dp, horizontal = 16.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(12.dp)
+    ) {
+        // 왼쪽 선
+        HorizontalDivider(
+            modifier = Modifier.weight(1f),
+            thickness = 1.dp,
+            color = color
+        )
+
+        // 중앙 메시지 (날짜 또는 "마지막으로 읽은 메시지")
+        Text(
+            text = message,
+            style = MaterialTheme.typography.labelMedium,
+            color = textColor
+        )
+
+        // 오른쪽 선
+        HorizontalDivider(
+            modifier = Modifier.weight(1f),
+            thickness = 1.dp,
+            color = color
+        )
+    }
+}
